Replace JpaAuditingConfiguration with OpenAPIConfig for secured REST API
In this commit, we removed JpaAuditingConfiguration, as it was not required anymore. On its lieu, OpenAPIConfig has been added to enable Swagger UI for secured REST API documentation.

SecurityConfig file was updated to disable CSRF and allow CORS configurations, considering security measures required for the OpenAPI integration.

Two new files, FermetureController and FermetureService, were added to handle 'Fermeture' endpoints. This includes get, create, and delete operations for 'Fermeture' resource.

"Dockerfile" has been added to ensure an organized process of building images, which will make it easier to create and manage the application using Docker.

Other minor changes include addition of new model Error to handle error messages, a ControllerAdvisor for exception handling, the FermetureServiceImpl to bridge the service with the repository, and an updated pom.xml to include the necessary dependencies.

Overall, this commit supports a secure and well-documented REST API with efficient error handling and docker compatibility.