Implement reservation status with email notifications
Added reservation status to the reservation functionality, including new methods in ReservationService and changes to related classes. The newly introduced ReservationStatus is an Enum used to track the status of a reservation, which can be 'accepted', 'waiting', or 'refused'.

Also introduced email notifications via the MailServiceImpl. Now, when a reservation is added or its status is changed, an email notification is sent to the user. Unit test for AuthService was also added. Spring Mail Starter dependency was included in pom.xml for mail service functionality.