V 1.0.0 Refactor ReservationService and improve date checks
A significant refactor of the `ReservationService` was implemented to enhance the checking of date availability in reservations. The entity `ReservationEntity` was modified to reflect changes in date terminology, improving readability and context understanding. Removal of `checkDateAvailable` function and replacement with `ensureDateIsAvailable` was done to throw exceptions when a date is not available, improving error handling. Test classes were updated to match these adjustments. Added date validation in `FermetureService` to prevent past or null dates.